+++ .github/workflows/deploy.yml
@@ -151,7 +151,7 @@ jobs:
         run: cd ./services/app-api && sls invoke -s $STAGE_PREFIX$branch_name -f migrate
       - name: Reset Test Data
         if: ${{ env.branch_name != 'production'}}
-        run: cd ./services/app-api && sls invoke -s $STAGE_PREFIX$branch_name -f resetData
+        run: cd ./services/admin && sls invoke -s $STAGE_PREFIX$branch_name -f resetData
   cypress-test:
     name: Cypress Tests
     needs: deploy
@@ -359,16 +359,3 @@ jobs:
           config: baseUrl=${{ env.APPLICATION_ENDPOINT }}
         env:
           SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
-  # reset:
-  #   name: Reset
-  #   needs:
-  #     - cypress-test
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #     - name: set branch_name
-  #       run: echo "branch_name=${GITHUB_REF#refs/heads/}" >> $GITHUB_ENV
-  #     - name: set stage_prefix
-  #       run: echo "STAGE_PREFIX=${{ secrets.STAGE_PREFIX }}" >> $STATE_PREFIX
-  #     - name: Reset Test Data
-  #       if: ${{ env.branch_name != 'production'}}
-  #       run: cd ./services/app-api && sls invoke -s $STAGE_PREFIX$branch_name -f resetData
