+++ services/admin/handlers/resetData.js
@@ -1,4 +1,3 @@
-import handler from "./libs/handler-lib";
 import AWS from "aws-sdk";
 
 const dynamoDb = new AWS.DynamoDB.DocumentClient(
