+++ services/app-api/serverless.yml
@@ -99,11 +99,6 @@ provider:
 
 functions:
 
-  resetData:
-    handler: resetData.main
-    role: LambdaApiRole
-    timeout: 180
-
   migrate:
     handler: migrate.main
     role: LambdaApiRole
